Docs: 스토리 ts 변경 및 중복 이름 수정
