+++ src/stories/Headerexample.stories.ts
@@ -1,0 +1,26 @@
+import type { Meta, StoryObj } from "@storybook/react";
+import { Headerexample } from "./Headerexample";
+
+const meta = {
+  title: "Example/Header",
+  component: Headerexample,
+  // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/writing-docs/autodocs
+  tags: ["autodocs"],
+  parameters: {
+    // More on how to position stories at: https://storybook.js.org/docs/configure/story-layout
+    layout: "fullscreen",
+  },
+} satisfies Meta<typeof Headerexample>;
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const LoggedIn: Story = {
+  args: {
+    user: {
+      name: "Jane Doe",
+    },
+  },
+};
+
+export const LoggedOut: Story = {};
